Refactor services
