+++ src/Authorize.hs
@@ -8,11 +8,12 @@ import qualified Data.Yaml as Y
 
 import System.IO (hFlush, stdout)
 
-import Web.Twitter.Conduit
+import Web.Twitter.Conduit hiding (map)
 import Web.Authenticate.OAuth as OA
 
 import App
 import Models
+import Types.Services
 
 authorize :: String -> IO ()
 authorize service = do
@@ -20,26 +21,30 @@ authorize service = do
     putStrLn ""
     putStr $ U.toString $ Y.encode auth
 
-getAuthorization :: String -> IO ServiceAuth
+getAuthorization :: String -> IO AppAuth
 getAuthorization "twitter" = authorizeTwitter
 getAuthorization _ = error "Invalid service name"
 
-authorizeTwitter :: IO ServiceAuth
-authorizeTwitter = do
-    app <- loadAppDefault
+withTwitter :: AppData -> (OAuth -> IO a) -> IO a
+withTwitter app act =
     case asTwitter (appServices app) of
         Nothing -> error "Twitter credentials not defined"
-        Just oauth -> do
-            mgr <- newManager tlsManagerSettings
-            tempCred <- OA.getTemporaryCredential oauth mgr
-            let url = OA.authorizeUrl oauth tempCred
-            pin <- getPIN url
-            cred <-
-                OA.getAccessToken
-                    oauth
-                    (OA.insert "oauth_verifier" pin tempCred)
-                    mgr
-            return $ twitterAuth cred
+        Just auth -> act auth
+
+authorizeTwitter :: IO AppAuth
+authorizeTwitter = do
+    app <- loadAppDefault
+    withTwitter app $ \auth -> do
+        mgr <- newManager tlsManagerSettings
+        tempCred <- OA.getTemporaryCredential auth mgr
+        let url = OA.authorizeUrl auth tempCred
+        pin <- getPIN url
+        cred <-
+            OA.getAccessToken
+                auth
+                (OA.insert "oauth_verifier" pin tempCred)
+                mgr
+        return $ toAppAuth $ twitterAuth cred
 
 getPIN :: String -> IO B.ByteString
 getPIN url = do
