+++ src/Types/Language.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
-module Language where
+module Types.Language where
 
 import Control.Monad
